requirements.txt
# Main API and Databricks integration
fastapi>=0.115
uvicorn[standard]>=0.30
databricks-sdk>=0.57
databricks-sql-connector>=3.5
databricks-agents>=0.16
mlflow>=3.6

# Data access and scraping
httpx>=0.27
requests>=2.32
beautifulsoup4>=4.12
rapidfuzz>=3.10
Unidecode>=1.3

# Models and configuration
pydantic>=2.8
python-dotenv>=1.0
cryptography>=43.0

# Local execution and tests
pyspark>=3.5
streamlit>=1.38
pytest>=8.3

# Optional: needed for the precursor Quebec ER agent notebooks kept in raw_export
databricks-langchain>=0.4
